refactor(examples): group examples into category subfolders; algod-backed NFT examples
Organize every example under examples/<category>/ — client, contract,
app, asset, transaction, account, logic_sig, wallet, nft — with explicit
[[example]] path entries. The example `name` is unchanged, so
`cargo run --example <name>` keeps working.

Move the NFT examples out of algonaut_nft/examples and into examples/nft,
and rewrite them against a live algod instead of mocking: real
suggested_params, submit, and real asset ids returned from the network
(no DemoParams stub, no hardcoded AssetId). Gated on
required-features = ["nft", "algod"]:
- nft_mint_arc3   — pure NFT, ARC-3 off-chain metadata
- nft_mint_arc19  — mutable NFT; mint then UpdateAsset the reserve
- nft_mint_arc69  — on-chain metadata in the acfg note
- nft_soulbound   — full ARC-71 issue/hold/revoke across issuer + funded holder
- nft_arc89_box   — mint for a real id, build the ARC-89 box offline
