+++ Cargo.toml
@@ -181,133 +181,201 @@ path = "tests/e2e/main.rs"
 test = false
 required-features = ["algod", "kmd"]
 
-# Examples declare the client features they exercise, so a slimmer build
-# (e.g. `--no-default-features --features algod`) skips the ones it cannot
-# compile instead of failing.
+# Examples are grouped by category under examples/<category>/. Each declares
+# the client features it exercises, so a slimmer build (e.g.
+# `--no-default-features --features algod`) skips the ones it cannot compile.
+# The `name` (used by `cargo run --example <name>`) is kept stable and unique.
+
+# --- client ---
 [[example]]
 name = "algod_client"
-required-features = ["algod"]
-
-[[example]]
-name = "method_call_dynamic"
-required-features = ["algod"]
-
-[[example]]
-name = "contract_arc4"
-required-features = ["algod"]
-
-[[example]]
-name = "contract_arc56"
-required-features = ["algod"]
-
-[[example]]
-name = "app_clear_state"
-required-features = ["algod"]
-
-[[example]]
-name = "app_close_out"
-required-features = ["algod"]
-
-[[example]]
-name = "app_create"
-required-features = ["algod"]
-
-[[example]]
-name = "app_delete"
-required-features = ["algod"]
-
-[[example]]
-name = "app_optin"
-required-features = ["algod"]
-
-[[example]]
-name = "app_update"
-required-features = ["algod"]
-
-[[example]]
-name = "asset_clawback"
-required-features = ["algod"]
-
-[[example]]
-name = "asset_create"
-required-features = ["algod"]
-
-[[example]]
-name = "asset_optin"
-required-features = ["algod"]
-
-[[example]]
-name = "asset_transfer"
-required-features = ["algod"]
-
-[[example]]
-name = "atomic"
-required-features = ["algod"]
-
-[[example]]
-name = "atomic_swap"
-required-features = ["algod"]
-
-[[example]]
-name = "indexer_query"
-required-features = ["indexer"]
-
-[[example]]
-name = "key_reg"
+path = "examples/client/algod_client.rs"
 required-features = ["algod"]
 
 [[example]]
 name = "kmd_client"
+path = "examples/client/kmd_client.rs"
 required-features = ["kmd"]
 
 [[example]]
-name = "kmd_sign_submit_transaction"
-required-features = ["algod", "kmd"]
+name = "indexer_query"
+path = "examples/client/indexer_query.rs"
+required-features = ["indexer"]
 
 [[example]]
 name = "last_block"
+path = "examples/client/last_block.rs"
+required-features = ["algod"]
+
+# --- contract ---
+[[example]]
+name = "contract_arc4"
+path = "examples/contract/arc4.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "logic_sig_contract_account"
+name = "contract_arc56"
+path = "examples/contract/arc56.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "logic_sig_delegated"
+name = "method_call_dynamic"
+path = "examples/contract/method_call_dynamic.rs"
+required-features = ["algod"]
+
+# --- app ---
+[[example]]
+name = "app_create"
+path = "examples/app/create.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "logic_sig_delegated_multi"
+name = "app_update"
+path = "examples/app/update.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "multi_sig"
+name = "app_delete"
+path = "examples/app/delete.rs"
 required-features = ["algod"]
 
+[[example]]
+name = "app_optin"
+path = "examples/app/optin.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "app_close_out"
+path = "examples/app/close_out.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "app_clear_state"
+path = "examples/app/clear_state.rs"
+required-features = ["algod"]
+
+# --- asset ---
+[[example]]
+name = "asset_create"
+path = "examples/asset/create.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "asset_transfer"
+path = "examples/asset/transfer.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "asset_clawback"
+path = "examples/asset/clawback.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "asset_optin"
+path = "examples/asset/optin.rs"
+required-features = ["algod"]
+
+# --- transaction ---
 [[example]]
 name = "payment"
+path = "examples/transaction/payment.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "quickstart"
-required-features = ["algod", "indexer", "kmd"]
-
-[[example]]
-name = "rekey"
+name = "atomic"
+path = "examples/transaction/atomic.rs"
 required-features = ["algod"]
 
 [[example]]
-name = "sign_offline"
+name = "atomic_swap"
+path = "examples/transaction/atomic_swap.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "key_reg"
+path = "examples/transaction/key_reg.rs"
 required-features = ["algod"]
 
 [[example]]
 name = "submit_file"
+path = "examples/transaction/submit_file.rs"
 required-features = ["algod"]
 
 [[example]]
+name = "quickstart"
+path = "examples/transaction/quickstart.rs"
+required-features = ["algod", "indexer", "kmd"]
+
+# --- account ---
+[[example]]
+name = "multi_sig"
+path = "examples/account/multi_sig.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "rekey"
+path = "examples/account/rekey.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "sign_offline"
+path = "examples/account/sign_offline.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "kmd_sign_submit_transaction"
+path = "examples/account/kmd_sign_submit_transaction.rs"
+required-features = ["algod", "kmd"]
+
+# --- logic_sig ---
+[[example]]
+name = "logic_sig_delegated"
+path = "examples/logic_sig/delegated.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "logic_sig_delegated_multi"
+path = "examples/logic_sig/delegated_multi.rs"
+required-features = ["algod"]
+
+[[example]]
+name = "logic_sig_contract_account"
+path = "examples/logic_sig/contract_account.rs"
+required-features = ["algod"]
+
+# --- wallet ---
+[[example]]
 name = "wallet_backup"
+path = "examples/wallet/backup.rs"
 required-features = ["kmd"]
 
 [[example]]
 name = "wallet_restore"
+path = "examples/wallet/restore.rs"
 required-features = ["kmd"]
+
+# --- nft (offline crate + algod for real suggested params, submit, asset ids) ---
+[[example]]
+name = "nft_mint_arc3"
+path = "examples/nft/mint_arc3.rs"
+required-features = ["nft", "algod"]
+
+[[example]]
+name = "nft_mint_arc19"
+path = "examples/nft/mint_arc19.rs"
+required-features = ["nft", "algod"]
+
+[[example]]
+name = "nft_mint_arc69"
+path = "examples/nft/mint_arc69.rs"
+required-features = ["nft", "algod"]
+
+[[example]]
+name = "nft_soulbound"
+path = "examples/nft/soulbound.rs"
+required-features = ["nft", "algod"]
+
+[[example]]
+name = "nft_arc89_box"
+path = "examples/nft/arc89_box.rs"
+required-features = ["nft", "algod"]
