+++ examples/account/kmd_sign_submit_transaction.rs
