+++ examples/app/clear_state.rs
