+++ examples/app/close_out.rs
