+++ examples/app/delete.rs
