+++ examples/app/optin.rs
