+++ examples/asset/clawback.rs
