+++ examples/asset/create.rs
