+++ examples/asset/optin.rs
