+++ examples/client/indexer_query.rs
