+++ examples/client/kmd_client.rs
