+++ examples/client/last_block.rs
