+++ examples/contract/arc4.rs
