+++ examples/contract/arc56.rs
