+++ examples/contract/method_call_dynamic.rs
