+++ examples/logic_sig/contract_account.rs
