+++ examples/logic_sig/delegated.rs
