+++ examples/logic_sig/delegated_multi.rs
