+++ examples/transaction/key_reg.rs
