+++ examples/wallet/backup.rs
