+++ examples/wallet/restore.rs
